+++ src/lib.rs
@@ -1,14 +1,22 @@
 //! Tibco EMS binding.
 
+#[cfg(feature = "ems-sys")]
 use enum_extract::extract;
+#[cfg(feature = "ems-sys")]
 use log::{error, trace};
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
+#[cfg(feature = "ems-sys")]
 use std::ffi::c_void;
+#[cfg(feature = "ems-sys")]
 use std::ffi::CStr;
+#[cfg(feature = "ems-sys")]
 use std::ffi::CString;
+use std::fmt;
 use std::io::Error;
+#[cfg(feature = "ems-sys")]
 use std::io::ErrorKind;
+#[cfg(feature = "ems-sys")]
 use std::ops::Deref;
 use std::sync::Arc;
 #[cfg(feature = "ems-sys")]
@@ -171,13 +179,13 @@ pub enum Message {
     ObjectMessage(ObjectMessage),
 }
 
-impl Message {
-    fn get_type(&self)->&str{
-        match self{
-            Message::TextMessage(_) => {"TextMessage"}
-            Message::BytesMessage(_) => {"BytesMessage"}
-            Message::MapMessage(_) => {"MapMessage"}
-            Message::ObjectMessage(_) => {"ObjectMessage"}
+impl fmt::Display for Message {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        match self {
+            Message::TextMessage(_) => write!(f, "TextMessage"),
+            Message::BytesMessage(_) => write!(f, "BytesMessage"),
+            Message::MapMessage(_) => write!(f, "MapMessage"),
+            Message::ObjectMessage(_) => write!(f, "ObjectMessage"),
         }
     }
 }
@@ -352,6 +360,7 @@ impl Connection {
         }
         Ok(session)
     }
+
     #[cfg(not(feature = "ems-sys"))]
     /// open a session
     pub fn transacted_session(&self) -> Result<Session, Error> {
@@ -360,6 +369,7 @@ impl Connection {
             producer_pointer: 0,
         })
     }
+
     #[cfg(feature = "ems-sys")]
     /// get active url from a ft connection
     /// this is only required for admin connections,
@@ -384,6 +394,7 @@ impl Connection {
             Ok(url.to_string())
         }
     }
+
     #[cfg(not(feature = "ems-sys"))]
     /// get active url from a ft connection
     /// this is only required for admin connections,
@@ -391,6 +402,7 @@ impl Connection {
     pub fn get_active_url(&self) -> Result<String, Error> {
         Ok("".to_string())
     }
+
     // open a consumer as stream of messages
     #[cfg(feature = "streaming")]
     pub fn open_stream<'stream, T: Into<Message>>(
@@ -418,7 +430,7 @@ impl Consumer {
     #[cfg(feature = "ems-sys")]
     /// receive messages from a consumer
     ///
-    /// function return after wait time with a Message or None
+    /// function returns after wait time with a Message or None
     /// a wait time of None blocks until a message is available
     pub fn receive_message(&self, wait_time_ms: Option<i64>) -> Result<Option<Message>, Error> {
         unsafe {
@@ -470,106 +482,111 @@ impl Consumer {
         }
     }
 
+    #[cfg(feature = "ems-sys")]
+    /// receive text messages from a consumer
+    ///
+    /// function returns after wait time with a Message or None
+    /// a wait time of None blocks until a message is available
     pub fn receive_text_message(
-        &self, wait_time_ms: Option<i64>
+        &self,
+        wait_time_ms: Option<i64>,
     ) -> Result<Option<TextMessage>, Error> {
         let msg_option = self.receive_message(wait_time_ms)?;
         match msg_option {
-            Some(msg)=>{
-                match &msg {
-                    Message::TextMessage(text_msg)=>{
-                        Ok(Some(text_msg.to_owned()))
-                    }
-                    _ => {
-                        Err(Error::new(
-                            ErrorKind::Other,
-                            format!("received message with unexpected type (expected: TextMessage, found: {}", msg.get_type()),
-                        ))
-                    }
-                }
-            }
-            None =>{
-                Ok(None)
-            }
+            Some(msg) => match &msg {
+                Message::TextMessage(text_msg) => Ok(Some(text_msg.to_owned())),
+                _ => Err(Error::new(
+                    ErrorKind::Other,
+                    format!(
+                        "received message with unexpected type (expected: TextMessage, found: {}",
+                        msg
+                    ),
+                )),
+            },
+            None => Ok(None),
         }
     }
 
-    pub fn receive_byte_message(
-        &self, wait_time_ms: Option<i64>
+    #[cfg(feature = "ems-sys")]
+    /// receive bytes messages from a consumer
+    ///
+    /// function returns after wait time with a Message or None
+    /// a wait time of None blocks until a message is available
+    pub fn receive_bytes_message(
+        &self,
+        wait_time_ms: Option<i64>,
     ) -> Result<Option<BytesMessage>, Error> {
         let msg_option = self.receive_message(wait_time_ms)?;
         match msg_option {
-            Some(msg)=>{
-                match &msg {
-                    Message::BytesMessage(bytes_msg)=>{
-                        Ok(Some(bytes_msg.to_owned()))
-                    }
-                    _ => {
-                        Err(Error::new(
-                            ErrorKind::Other,
-                            format!("received message with unexpected type (expected: BytesMessage, found: {}", msg.get_type()),
-                        ))
-                    }
-                }
-            }
-            None =>{
-                Ok(None)
-            }
+            Some(msg) => match &msg {
+                Message::BytesMessage(bytes_msg) => Ok(Some(bytes_msg.to_owned())),
+                _ => Err(Error::new(
+                    ErrorKind::Other,
+                    format!(
+                        "received message with unexpected type (expected: BytesMessage, found: {}",
+                        msg
+                    ),
+                )),
+            },
+            None => Ok(None),
         }
     }
 
+    #[cfg(feature = "ems-sys")]
+    /// receive map messages from a consumer
+    ///
+    /// function returns after wait time with a Message or None
+    /// a wait time of None blocks until a message is available
     pub fn receive_map_message(
-        &self, wait_time_ms: Option<i64>
+        &self,
+        wait_time_ms: Option<i64>,
     ) -> Result<Option<MapMessage>, Error> {
         let msg_option = self.receive_message(wait_time_ms)?;
         match msg_option {
-            Some(msg)=>{
-                match &msg {
-                    Message::MapMessage(map_msg)=>{
-                        Ok(Some(map_msg.to_owned()))
-                    }
-                    _ => {
-                        Err(Error::new(
-                            ErrorKind::Other,
-                            format!("received message with unexpected type (expected: MapMessage, found: {}", msg.get_type()),
-                        ))
-                    }
-                }
-            }
-            None =>{
-                Ok(None)
-            }
+            Some(msg) => match &msg {
+                Message::MapMessage(map_msg) => Ok(Some(map_msg.to_owned())),
+                _ => Err(Error::new(
+                    ErrorKind::Other,
+                    format!(
+                        "received message with unexpected type (expected: MapMessage, found: {}",
+                        msg
+                    ),
+                )),
+            },
+            None => Ok(None),
         }
     }
 
+    #[cfg(feature = "ems-sys")]
+    /// receive object messages from a consumer
+    ///
+    /// function returns after wait time with a Message or None
+    /// a wait time of None blocks until a message is available
     pub fn receive_object_message(
-        &self, wait_time_ms: Option<i64>
+        &self,
+        wait_time_ms: Option<i64>,
     ) -> Result<Option<ObjectMessage>, Error> {
         let msg_option = self.receive_message(wait_time_ms)?;
         match msg_option {
-            Some(msg)=>{
-                match &msg {
-                    Message::ObjectMessage(object_msg)=>{
-                        Ok(Some(object_msg.to_owned()))
-                    }
-                    _ => {
-                        Err(Error::new(
-                            ErrorKind::Other,
-                            format!("received message with unexpected type (expected: ObjectMessage, found: {}", msg.get_type()),
-                        ))
-                    }
-                }
-            }
-            None =>{
-                Ok(None)
-            }
+            Some(msg) => match &msg {
+                Message::ObjectMessage(object_msg) => Ok(Some(object_msg.to_owned())),
+                _ => Err(Error::new(
+                    ErrorKind::Other,
+                    format!(
+                        "received message with unexpected type (expected: ObjectMessage, found: {}",
+                        msg
+                    ),
+                )),
+            },
+            None => Ok(None),
         }
     }
+
     #[cfg(not(feature = "ems-sys"))]
     /// receive messages from a consumer
     ///
-    /// function return after wait time with a Message or None
-    /// a wait time of None blocks until a message is available
+    /// function returns after wait time with a Message or None
+    /// the wait time is ignored
     pub fn receive_message(&self, _wait_time_ms: Option<i64>) -> Result<Option<Message>, Error> {
         unsafe {
             let consumer_destination = SERVER.consumer.clone().unwrap();
@@ -578,6 +595,101 @@ impl Consumer {
                 if dest == consumer_destination {
                     return Ok(Some(msg));
                 }
+            }
+        }
+        Ok(None)
+    }
+    #[cfg(not(feature = "ems-sys"))]
+    /// receive text messages from a consumer
+    ///
+    /// function returns after wait time with a Message or None
+    /// the wait time is ignored
+    pub fn receive_text_message(
+        &self,
+        _wait_time_ms: Option<i64>,
+    ) -> Result<Option<TextMessage>, Error> {
+        unsafe {
+            let consumer_destination = SERVER.consumer.clone().unwrap();
+            let messages = SERVER.messages.clone();
+            for (dest, msg) in messages {
+                if dest == consumer_destination {
+                    return match msg {
+                        Message::TextMessage(ref msg) => Ok(Some(msg.to_owned())),
+                        _ => Ok(None),
+                    };
+                };
+            }
+        }
+        Ok(None)
+    }
+
+    #[cfg(not(feature = "ems-sys"))]
+    /// receive bytes messages from a consumer
+    ///
+    /// function returns after wait time with a Message or None
+    /// the wait time is ignored
+    pub fn receive_bytes_message(
+        &self,
+        _wait_time_ms: Option<i64>,
+    ) -> Result<Option<BytesMessage>, Error> {
+        unsafe {
+            let consumer_destination = SERVER.consumer.clone().unwrap();
+            let messages = SERVER.messages.clone();
+            for (dest, msg) in messages {
+                if dest == consumer_destination {
+                    return match msg {
+                        Message::BytesMessage(ref msg) => Ok(Some(msg.to_owned())),
+                        _ => Ok(None),
+                    };
+                };
+            }
+        }
+        Ok(None)
+    }
+
+    #[cfg(not(feature = "ems-sys"))]
+    /// receive map messages from a consumer
+    ///
+    /// function returns after wait time with a Message or None
+    /// the wait time is ignored
+    pub fn receive_map_message(
+        &self,
+        _wait_time_ms: Option<i64>,
+    ) -> Result<Option<MapMessage>, Error> {
+        unsafe {
+            let consumer_destination = SERVER.consumer.clone().unwrap();
+            let messages = SERVER.messages.clone();
+            for (dest, msg) in messages {
+                if dest == consumer_destination {
+                    return match msg {
+                        Message::MapMessage(ref msg) => Ok(Some(msg.to_owned())),
+                        _ => Ok(None),
+                    };
+                };
+            }
+        }
+        Ok(None)
+    }
+
+    #[cfg(not(feature = "ems-sys"))]
+    /// receive object messages from a consumer
+    ///
+    /// function returns after wait time with a Message or None
+    /// the wait time is ignored
+    pub fn receive_object_message(
+        &self,
+        _wait_time_ms: Option<i64>,
+    ) -> Result<Option<ObjectMessage>, Error> {
+        unsafe {
+            let consumer_destination = SERVER.consumer.clone().unwrap();
+            let messages = SERVER.messages.clone();
+            for (dest, msg) in messages {
+                if dest == consumer_destination {
+                    return match msg {
+                        Message::ObjectMessage(ref msg) => Ok(Some(msg.to_owned())),
+                        _ => Ok(None),
+                    };
+                };
             }
         }
         Ok(None)
@@ -766,9 +878,9 @@ impl Session {
     /// open a message consumer for a topic
     pub fn topic_consumer(
         &self,
-        destination: &Destination,
-        subscription_name: &str,
-        selector: Option<&str>,
+        _destination: &Destination,
+        _subscription_name: &str,
+        _selector: Option<&str>,
     ) -> Result<Consumer, Error> {
         unimplemented!()
     }
@@ -852,9 +964,9 @@ impl Session {
     /// open a durable message consumer for a topic
     pub fn topic_durable_consumer(
         &self,
-        destination: &Destination,
-        durable_name: &str,
-        selector: Option<&str>,
+        _destination: &Destination,
+        _durable_name: &str,
+        _selector: Option<&str>,
     ) -> Result<Consumer, Error> {
         unimplemented!()
     }
@@ -882,6 +994,34 @@ impl Session {
     /// close a session
     fn close(&self) {}
 
+    #[cfg(feature = "tracing")]
+    fn add_trace_to_message(&self, message: &mut Message) -> impl opentelemetry::trace::Span {
+        let tracer_provider = opentelemetry::global::tracer_provider();
+        use opentelemetry::trace::Span;
+        use opentelemetry::trace::Tracer;
+        use opentelemetry::trace::TracerProvider;
+        let tracer = tracer_provider.versioned_tracer("ems", Some("0.5"), None);
+        let span = tracer.start("send");
+        let headers = match message {
+            Message::BytesMessage(b) => b.header.as_mut(),
+            Message::MapMessage(m) => m.header.as_mut(),
+            Message::TextMessage(t) => t.header.as_mut(),
+            Message::ObjectMessage(o) => o.header.as_mut(),
+        };
+        let ctx = span.span_context();
+        if let Some(e) = headers {
+            e.insert(
+                "spanId".to_string(),
+                TypedValue::String(ctx.span_id().to_string()),
+            );
+            e.insert(
+                "traceId".to_string(),
+                TypedValue::String(ctx.trace_id().to_string()),
+            );
+        };
+        span
+    }
+
     #[cfg(feature = "ems-sys")]
     /// sending a message to a destination (only queues are supported)
     pub fn send_message<M: Into<Message>>(
@@ -889,12 +1029,24 @@ impl Session {
         destination: &Destination,
         message: M,
     ) -> Result<(), Error> {
+        #[cfg(feature = "tracing")]
+        let mut message: Message = message.into();
+        #[cfg(not(feature = "tracing"))]
         let message: Message = message.into();
+        
         let mut dest: usize = 0;
         let mut local_producer: usize = 0;
+        #[cfg(feature = "tracing")]
+        let mut span = self.add_trace_to_message(&mut message);
+        #[cfg(feature = "tracing")]
+        use opentelemetry::trace::Span;
         unsafe {
             match destination {
                 Destination::Queue(name) => {
+                    #[cfg(feature = "tracing")]
+                    span.set_attribute(opentelemetry::KeyValue::new("messaging.destination", name.clone()));
+                    #[cfg(feature = "tracing")]
+                    span.set_attribute(opentelemetry::KeyValue::new("messaging.destination_kind", "queue"));
                     let c_destination = CString::new(name.clone()).unwrap();
                     let status = tibco_ems_sys::tibemsDestination_Create(
                         &mut dest,
@@ -995,6 +1147,8 @@ impl Session {
                 _ => error!("tibemsDestination_Destroy: {:?}", status),
             }
         }
+        #[cfg(feature = "tracing")]
+        span.end();
         Ok(())
     }
 
@@ -1177,12 +1331,12 @@ impl Session {
     }
 
     #[cfg(not(feature = "ems-sys"))]
-    /// request/reply - always returns timeout
+    /// request/reply - always returns none
     pub fn request_reply<M: Into<Message>>(
         &self,
         destination: &Destination,
         message: M,
-        timeout: i64,
+        _timeout: i64,
     ) -> Result<Option<Message>, Error> {
         let message: Message = message.into();
         unsafe {
@@ -1246,6 +1400,21 @@ pub enum TypedValue {
     Map(MapMessage),
     /// represents a boolean Value
     Boolean(bool),
+}
+
+impl fmt::Display for TypedValue {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        match self {
+            TypedValue::String(s) => write!(f, "{s}"),
+            TypedValue::Boolean(b) => write!(f, "{b}"),
+            TypedValue::Integer(b) => write!(f, "{b}"),
+            TypedValue::Long(b) => write!(f, "{b}"),
+            TypedValue::Float(b) => write!(f, "{b}"),
+            TypedValue::Double(b) => write!(f, "{b}"),
+            TypedValue::Binary(b) => write!(f, "{b:?}"),
+            TypedValue::Map(m) => write!(f, "{m:?}"),
+        }
+    }
 }
 
 impl Message {
@@ -1563,13 +1732,10 @@ fn build_message_pointer_from_message(message: &Message) -> usize {
             }
             //look for jms type
             if let Some(jms_type) = headers.get("JMSType") {
-                let jms_type_val = extract!(TypedValue::String(_), jms_type)
-                    .expect("extract correlation id");
+                let jms_type_val =
+                    extract!(TypedValue::String(_), jms_type).expect("extract correlation id");
                 let c_jms_type = CString::new(jms_type_val.as_str()).unwrap();
-                let status = tibco_ems_sys::tibemsMsg_SetType(
-                    msg_pointer,
-                    c_jms_type.as_ptr(),
-                );
+                let status = tibco_ems_sys::tibemsMsg_SetType(msg_pointer, c_jms_type.as_ptr());
                 match status {
                     tibems_status::TIBEMS_OK => trace!("tibemsMsg_SetType: {:?}", status),
                     _ => error!("tibemsMsg_SetType: {:?}", status),